requirements.txt
streamlit==1.32.0
python-docx==1.1.0
requests==2.31.0
beautifulsoup4==4.12.3 